Redefine postal code verification methods to use setting config parameter
